+++ src/cli.js
@@ -1,37 +1,17 @@
-import Command from './parser/command';
-import Context from './parser/context';
-import debug from './lib/debug';
-import E from './lib/errors';
-import Extension from './parser/extension';
-import helpCommand, { renderHelp } from './commands/help';
-import Parser from './parser/parser';
+import Command from './parser/command.js';
+import Context from './parser/context.js';
+import debug from './lib/debug.js';
+import E from './lib/errors.js';
+import Extension from './parser/extension.js';
+import helpCommand, { renderHelp } from './commands/help.js';
+import Parser from './parser/parser.js';
 import pluralize from 'pluralize';
-import Terminal from './terminal';
-import WebSocket, { Server as WebSocketServer } from 'ws';
+import Terminal from './terminal.js';
 
-import * as ansi from './lib/ansi';
-
-import { declareCLIKitClass, decode, split, assertNodeJSVersion } from './lib/util';
-import { EventEmitter } from 'events';
-import { generateKey } from './lib/keys';
-import { WriteStream } from 'tty';
+import { assertNodeJSVersion, declareCLIKitClass } from './lib/util.js';
 
 const { error, log, warn } = debug('cli-kit:cli');
 const { highlight, note }  = debug.styles;
-
-/**
- * Writes data to a websocket.
- */
-class OutputSocket extends WriteStream {
-	constructor(fd, ws) {
-		super(fd);
-		this.ws = ws;
-	}
-
-	write(chunk) {
-		this.ws.send(chunk.replace(/(?<!\r)\n/g, '\r\n'));
-	}
-}
 
 /**
  * Defines a CLI context and is responsible for parsing the command line arguments.
@@ -71,12 +51,6 @@ export default class CLI extends Context {
 	 * @param {String} [params.name] - The name of the program. If not set, defaults to `"program"`
 	 * in the help outut and `"This application"` in the Node version assertion.
 	 * @param {String} [params.nodeVersion] - The required Node.js version to run the app.
-	 * @param {Boolean} [opts.serverMode=false] - When `true`, makes things such that `exec()`
-	 * doesn't change any global state by deep cloning the entire context tree every time the
-	 * arguments are parsed and changing the process state. Enabling this only makes sense if the
-	 * `CLI` instance is going to be reused to parse arguments multiple times and if the state of
-	 * the context tree is going to be modified during parsing (i.e. via a callback). It also
-	 * prevents ctrl-c (SIGINT) and does not set an exit code.
 	 * @param {Boolean} [params.showBannerForExternalCLIs=false] - If `true`, shows the `CLI`
 	 * banner, assuming banner is enabled, for non-cli-kit enabled CLIs.
 	 * @param {Boolean} [params.showHelpOnError=true] - If an error occurs and `help` is enabled,
@@ -143,7 +117,6 @@ export default class CLI extends Context {
 		this.hideNoBannerOption        = params.hideNoBannerOption;
 		this.hideNoColorOption         = params.hideNoColorOption;
 		this.nodeVersion               = params.nodeVersion;
-		this.serverMode                = params.serverMode;
 		this.showBannerForExternalCLIs = params.showBannerForExternalCLIs;
 		this.showHelpOnError           = params.showHelpOnError;
 		this.styles                    = Object.assign({}, debug.styles, params.styles);
@@ -151,10 +124,7 @@ export default class CLI extends Context {
 		this.version                   = params.version;
 		this.warnings                  = [];
 
-		// if we're not in server mode, wire up the SIGINT handler
-		if (!this.serverMode) {
-			this.terminal.on('SIGINT', () => process.kill(process.pid, 'SIGINT'));
-		}
+		this.terminal.on('SIGINT', () => process.kill(process.pid, 'SIGINT'));
 
 		// add the built-in help
 		if (this.help) {
@@ -211,93 +181,6 @@ export default class CLI extends Context {
 				}
 			}
 		}
-	}
-
-	/**
-	 * Connects to a cli-kit WebSocket server and initializes a terminal session.
-	 *
-	 * @param {String} url - The URL to connect to.
-	 * @param {Object} [opts] - Various options.
-	 * @param {Object} [opts.headers] - HTTP headers to send when creating the WebSocket.
-	 * @param {Termianl} [opts.terminal] - A terminal instance to override the default CLI terminal
-	 * instance.
-	 * @param {Number} [opts.timeout=5000] - The number of milliseconds to wait to connect to the
-	 * server and complete the initialization handshake.
-	 * @returns {Promise<Object>} Resolves an `EventEmitter` based handle containing a `send()`
-	 * function to send data as if from `stdin`.
-	 * @access public
-	 */
-	static async connect(url, opts = {}) {
-		if (!url || typeof url !== 'string') {
-			throw E.INVALID_ARGUMENT('Expected URL to be a string', { name: 'url', scope: 'CLI.connect', value: url });
-		}
-
-		if (!opts || typeof opts !== 'object') {
-			throw E.INVALID_ARGUMENT('Expected options to be an object', { name: 'opts', scope: 'CLI.connect', value: opts });
-		}
-
-		let term = opts.terminal;
-		delete opts.terminal;
-
-		if (!term) {
-			term = new Terminal();
-		} else if (!(term instanceof Terminal)) {
-			throw E.INVALID_ARGUMENT('Expected terminal to be a Terminal instance', { name: 'opts.terminal', scope: 'CLI.connect', value: term });
-		}
-
-		return await new Promise((resolve, reject) => {
-			log(`Connecting to ${highlight(url)}`);
-			const ws = new WebSocket(url, opts);
-			ws.binaryType = 'arraybuffer';
-
-			const handle = new EventEmitter();
-			handle.send = chunk => {
-				if (ws.readyState === 1) {
-					ws.send(chunk);
-				}
-			};
-
-			const initTimer = setTimeout(() => {
-				const err = new Error(ws.readyState === 1 ? 'Failed to initialize terminal session' : 'Failed to connect to server');
-				err.code = 'ETIMEOUT';
-				reject(err);
-			}, opts.timeout || 5000);
-
-			ws.on('close', (code, reason) => handle.emit('close', code, reason));
-
-			ws.on('error', err => handle.emit('error', err));
-
-			ws.on('message', msg => {
-				if (msg === ansi.cursor.get && ws.readyState === 1) {
-					clearTimeout(initTimer);
-					ws.send(`${ansi.esc}${term.rows};${term.columns}R`);
-					resolve(handle);
-					return;
-				}
-
-				const code = msg.match(ansi.custom.exit.re);
-				if (code) {
-					handle.emit('exit', code[1]);
-					return;
-				}
-
-				term.stdout.write(msg);
-			});
-
-			term.on('keypress', (chunk, key) => {
-				// console.log('KEYPRESS', chunk === undefined ? chunk : Buffer.from(chunk), key, Buffer.from(key.sequence));
-				handle.send(ansi.custom.keypress(key));
-			});
-
-			term.on('resize', ({ rows, columns }) => {
-				handle.send(`${ansi.esc}${rows};${columns}R`);
-			});
-
-			term.on('SIGINT', () => {
-				ws.close();
-				process.exit();
-			});
-		});
 	}
 
 	/**
@@ -452,9 +335,9 @@ export default class CLI extends Context {
 		};
 
 		try {
-			const cli = this.serverMode ? new Context().init(this, true) : this;
+			const cli = this;
 
-			log(`Parsing ${__argv.length} argument${__argv.length !== 1 ? 's' : ''} ${note(`(server mode: ${!!this.serverMode})`)}`);
+			log(`Parsing ${__argv.length} argument${__argv.length !== 1 ? 's' : ''}`);
 
 			// parse the command line arguments
 			const {
@@ -578,15 +461,11 @@ export default class CLI extends Context {
 				})(results);
 			}
 
-			if (!this.serverMode) {
-				process.exitCode = results.exitCode();
-			}
+			process.exitCode = results.exitCode();
 
 			return results;
 		} catch (err) {
-			if (!this.serverMode) {
-				error(err.stack || err.message || err.toString() || 'Unknown error');
-			}
+			error(err.stack || err.message || err.toString() || 'Unknown error');
 
 			if (err.json === undefined && results.cmd?.prop('jsonMode')) {
 				err.json = true;
@@ -606,248 +485,5 @@ export default class CLI extends Context {
 
 			throw err;
 		}
-	}
-
-	/**
-	 * Starts a WebSocket server.
-	 *
-	 * @param {Object} [opts] - WebSocket server options. Visit
-	 * https://github.com/websockets/ws/blob/HEAD/doc/ws.md#new-websocketserveroptions-callback for
-	 * more information.
-	 * @returns {Promise<WebSocketServer>}
-	 * @access public
-	 */
-	async listen(opts = {}) {
-		if (!opts || typeof opts !== 'object') {
-			throw E.INVALID_ARGUMENT('Expected options to be an object', { name: 'opts', scope: 'CLI.listen', value: opts });
-		}
-
-		if (opts.port !== undefined && (typeof opts.port !== 'number' || opts.port < 1 || opts.port > 65535)) {
-			throw E.INVALID_ARGUMENT('Expected port to be a number between 1 and 65535', { name: 'opts.port', scope: 'CLI.listen', value: opts.port });
-		}
-
-		log(`Starting WebSocketServer${opts.port ? ` on port ${highlight(opts.port)}` : ''}...`);
-
-		if (!this.serverMode) {
-			log('Enabling server mode');
-			this.serverMode = true;
-		}
-
-		return new Promise((resolve, reject) => {
-			this.server = new WebSocketServer(opts, () => resolve(this.server));
-
-			this.server.on('connection', (ws, req) => {
-				const { remoteAddress, remotePort } = req.socket;
-				const key = remoteAddress + ':' + remotePort;
-
-				try {
-					const { headers } = req;
-					const terminal = new Terminal({
-						stdout: new OutputSocket(1, ws),
-						stderr: new OutputSocket(2, ws)
-					});
-					let buffer = '';
-					let current = null;
-					let echo = false;
-
-					log(`${highlight(key)} upgraded to WebSocket`);
-					log(headers);
-
-					ws.on('close', () => log(`${highlight(key)} closed WebSocket`));
-
-					ws.on('error', err => {
-						if (err.code !== 'ECONNRESET') {
-							error(err);
-						}
-					});
-
-					const exec = async (args, post) => {
-						const command = split(args);
-						log(`Running: ${highlight(command)}`);
-
-						current = this.exec(command, {
-							data: {
-								cwd:       decode(headers['clikit-cwd']),
-								env:       decode(headers['clikit-env']),
-								userAgent: headers['user-agent'] || undefined
-							},
-							parentContextNames: decode(headers['clikit-parents']),
-							terminal
-						});
-
-						if (typeof post === 'function') {
-							post();
-						}
-
-						let ec = 1;
-
-						try {
-							const { exitCode } = await current;
-							ec = exitCode() || 0;
-						} catch (err) {
-							error(err.stack || err.message || err.toString());
-							terminal.stderr.write(`${this.styles.error(err.toString())}\n`);
-						} finally {
-							current = null;
-							log(`Command finished (code ${ec})`);
-							ws.send(ansi.custom.exit(ec));
-						}
-					};
-
-					ws.on('message', async msg => {
-						if (Buffer.isBuffer(msg)) {
-							msg = msg.toString();
-						}
-
-						let m;
-
-						try {
-							// check if we received a cursor message
-							m = msg.match(ansi.cursor.position);
-							if (m) {
-								const rows = terminal.stdout.rows = terminal.stderr.rows = ~~m[1];
-								const cols = terminal.stdout.columns = terminal.stderr.columns = ~~m[2];
-								log(`Terminal set to ${highlight(cols)} x ${highlight(rows)}`);
-								return;
-							}
-
-							// check if we received an echo message
-							m = msg.match(ansi.custom.echo.re);
-							if (m) {
-								echo = m[1] !== 'off';
-								log(`Setting echo ${highlight(echo ? 'on' : 'off')}`);
-								return;
-							}
-
-							// check if we received an execute message
-							m = msg.match(ansi.custom.exec.re);
-							if (m) {
-								return await exec(decode(m[1]));
-							}
-
-							// check if we received a keypress message
-							m = msg.match(ansi.custom.keypress.re);
-							if (m) {
-								const key = decode(m[1]);
-
-								if (current) {
-									terminal.stdin.emit('keypress', key.sequence, key);
-									return;
-								}
-
-								msg = key.sequence;
-								m = null;
-							}
-
-							// message is a raw message, so we have to clean it up, buffer, and manually dispatch
-
-							msg = msg.replace(/\r\n|\n|\r/g, '\r\n'); // normalize new lines
-							msg = msg.replace(/\x7f/g, '\b'); // normalize backspaces
-
-							// TODO: support cursor position
-
-							// repeat back to the client what they just passed us
-							if (echo) {
-								ws.send(msg.replace(/[\b]/g, '\b \b'));
-							}
-
-							// if there's already an active command, treat message as an incoming key from stdin
-							if (current) {
-								terminal.stdin.emit('keypress', msg, generateKey(msg));
-								return;
-							}
-
-							// no pending command, so we need to buffer and as soon as we see a line return,
-							// then we execute it and any remaining characters are treated as keypresses
-
-							buffer += msg;
-
-							// replace backspaces
-							for (let p = 0; (p = buffer.indexOf('\b', p)) !== -1;) {
-								buffer = buffer.substring(0, p - 1) + buffer.substring(p + 1);
-							}
-
-							let p = buffer.indexOf('\r');
-							if (p !== -1) {
-								const command = buffer.substring(0, p);
-								const str = buffer.substring(p + 2);
-								buffer = '';
-								await exec(command, () => {
-									if (str.length) {
-										try {
-											terminal.stdin.write(str);
-										} catch (err) {
-											warn('Failed to write to stdin');
-											warn(err);
-										}
-									}
-								});
-							}
-						} finally {
-							if (m) {
-								buffer = '';
-							}
-						}
-					});
-
-					// get the remote terminal size
-					ws.send(ansi.cursor.save);
-					ws.send(ansi.cursor.move(999, 999));
-					ws.send(ansi.cursor.get);
-					ws.send(ansi.cursor.restore);
-				} catch (err) {
-					warn(err);
-					warn(`Hanging up ${highlight(key)}`);
-				}
-			});
-
-			this.server.on('error', reject);
-		});
-	}
-
-	/**
-	 * Returns the schema for the CLI and all child contexts.
-	 *
-	 * @param {Object} [opts] - Various options.
-	 * @param {Object} [opts.data] - User-defined data to pass into the selected command.
-	 * @returns {Object}
-	 * @access public
-	 */
-	async schema(opts = {}) {
-		const obj = {
-			args:       [],
-			banner:     String(typeof this.banner === 'function' ? (await this.banner(opts)) : this.banner).trim(),
-			commands:   {},
-			desc:       this.desc,
-			extensions: {},
-			name:       this.name,
-			options:    {},
-			title:      this.title,
-			version:    String(typeof this.version === 'function' ? (await this.version(opts)) : this.version).trim()
-		};
-
-		for (const arg of this.args) {
-			if (!arg.hidden) {
-				obj.args.push(arg.schema);
-			}
-		}
-
-		for (const [ name, cmd ] of this.commands.entries()) {
-			obj.commands[name] = await cmd.schema(opts);
-		}
-
-		for (const [ name, ext ] of this.extensions.entries()) {
-			obj.extensions[name] = await ext.schema(opts);
-		}
-
-		for (const options of this.options.values()) {
-			for (const opt of options) {
-				if (!opt.hidden) {
-					obj.options[opt.format] = await opt.schema(opts);
-				}
-			}
-		}
-
-		return obj;
 	}
 }