+++ src/lib/util.js
@@ -1,10 +1,13 @@
 import argvSplit from 'argv-split';
 import fs from 'fs-extra';
-import E from './errors';
+import E from './errors.js';
 import path from 'path';
-import pkgDir from 'pkg-dir';
 import semver from 'semver';
 import which from 'which';
+import { fileURLToPath } from 'url';
+import { packageDirectorySync } from 'pkg-dir';
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 /**
  * The required Node.js version for cli-kit. This is used to assert the Node version at runtime.
@@ -99,7 +102,8 @@ export function findPackage(searchPath) {
 	let clikit = false;
 	let json = {};
 	let main = null;
-	let root = pkgDir.sync(searchPath) || null;
+	let root = packageDirectorySync({ cwd: searchPath }) || null;
+	let esm = false;
 
 	// don't let the tests think they are cli-kit
 	if (root === path.resolve(__dirname, '..', '..')) {
@@ -131,6 +135,9 @@ export function findPackage(searchPath) {
 			throw E.INVALID_PACKAGE_JSON('Invalid package.json: expected object', { file, name: 'package.json.invalid', scope: 'util.findPackage', value: json });
 		}
 
+		// detect if we have an ES module
+		esm = json.type === 'module';
+
 		if (json.clikit || json['cli-kit']) {
 			clikit = true;
 			Object.assign(json, json.clikit, json['cli-kit']);
@@ -158,7 +165,7 @@ export function findPackage(searchPath) {
 		root = main ? path.dirname(main) : null;
 	}
 
-	return { clikit, json, main, root };
+	return { clikit, esm, json, main, root };
 }
 
 /**
@@ -197,8 +204,6 @@ export function isFile(file) {
 	}
 	return false;
 }
-
-export { pkgDir };
 
 /**
  * Splits an argv (argument vector) string.