+++ src/parser/command.js
@@ -1,11 +1,10 @@
-import Context from './context';
-import debug from '../lib/debug';
-import E from '../lib/errors';
+import Context from './context.js';
+import debug from '../lib/debug.js';
+import E from '../lib/errors.js';
 import fs from 'fs';
-import helpCommand from '../commands/help';
+import helpCommand from '../commands/help.js';
 import path from 'path';
-
-import { declareCLIKitClass } from '../lib/util';
+import { declareCLIKitClass } from '../lib/util.js';
 
 const { log } = debug('cli-kit:command');
 const { highlight } = debug.styles;
@@ -58,29 +57,22 @@ export default class Command extends Context {
 	 *   new Command(new Command('foo'))
 	 */
 	constructor(name, params = {}) {
+		super(null); // null will skip the init since we do it applyParams()
+		declareCLIKitClass(this, 'Command');
+		this.applyParams(name, params);
+	}
+
+	/**
+	 * Initializes the command.
+	 *
+	 * @param {String} name - The command name.
+	 * @param {Object} params - Various parameters.
+	 * @access private
+	 */
+	applyParams(name, params) {
 		if (name && typeof name === 'string' && path.isAbsolute(name) && fs.existsSync(name)) {
-			let ctx;
-			try {
-				log(`Requiring ${highlight(name)}`);
-				ctx = require(name);
-				if (!ctx || typeof ctx !== 'object') {
-					throw new Error('Command must export an object');
-				}
-
-				// if this is an ES6 module, grab the default export
-				if (ctx.__esModule) {
-					ctx = ctx.default;
-				}
-
-				if (!ctx || typeof ctx !== 'object') {
-					throw new Error('Command must export an object');
-				}
-
-				name = ctx.name || path.parse(name).name;
-				params = ctx;
-			} catch (err) {
-				throw E.INVALID_COMMAND(`Bad command "${name}": ${err.message}`, { name: name, scope: 'Command.constructor', value: err });
-			}
+			params.modulePath = name;
+			name = path.parse(name).name;
 		}
 
 		if (!name || typeof name !== 'string') {
@@ -116,7 +108,7 @@ export default class Command extends Context {
 		}
 
 		if (!name) {
-			throw  E.INVALID_ARGUMENT('Expected command name format to contain at least one non-aliased name', { name: 'format', scope: 'Command.constructor', value: format });
+			throw E.INVALID_ARGUMENT('Expected command name format to contain at least one non-aliased name', { name: 'format', scope: 'Command.constructor', value: format });
 		}
 
 		if (!params || (typeof params !== 'object' || Array.isArray(params))) {
@@ -172,8 +164,7 @@ export default class Command extends Context {
 			params.args = params.args ? [ ...args, ...params.args ] : args;
 		}
 
-		super(params);
-		declareCLIKitClass(this, 'Command');
+		this.init(params);
 
 		if (params.action) {
 			this.action = params.action;
@@ -187,9 +178,11 @@ export default class Command extends Context {
 		this._aliases       = this.createAliases(aliases, params.aliases);
 		this.callback       = params.callback;
 		this.clikitHelp     = params.clikitHelp;
-		this.help           = params.help || {};
 		this.defaultCommand = params.defaultCommand;
+		this.help           = params.help || {};
 		this.hidden         = !!params.hidden;
+		this.loaded         = !params.modulePath;
+		this.modulePath     = params.modulePath;
 
 		// mix in any other custom props
 		for (const [ key, value ] of Object.entries(params)) {
@@ -294,6 +287,40 @@ export default class Command extends Context {
 			}
 		} else {
 			this._help = {};
+		}
+	}
+
+	/**
+	 * Loads this command if it is defined in an external file.
+	 *
+	 * @returns {Promise}
+	 * @access public
+	 */
+	async load() {
+		if (this.loaded) {
+			return;
+		}
+
+		try {
+			log(`Importing ${highlight(this.modulePath)}`);
+			let ctx = await import(`file://${this.modulePath}`);
+			if (!ctx || typeof ctx !== 'object') {
+				throw new Error('Command must export an object');
+			}
+
+			// if this is an ES6 module, grab the default export
+			if (ctx.default) {
+				ctx = ctx.default;
+			}
+
+			if (!ctx || typeof ctx !== 'object') {
+				throw new Error('Command must export an object');
+			}
+
+			this.applyParams(ctx.name || this.name, ctx);
+			this.loaded = true;
+		} catch (err) {
+			throw E.INVALID_COMMAND(`Bad command "${this.name}": ${err.message}`, { name: this.name, scope: 'Command.load', value: err });
 		}
 	}