+++ src/parser/lookup.js
@@ -1,4 +1,4 @@
-import debug from '../lib/debug';
+import debug from '../lib/debug.js';
 
 const { highlight, note } = debug.styles;
 
@@ -55,7 +55,8 @@ export default class Lookup {
 		if (Object.keys(this.commands).length) {
 			lines.push(note('  Commands:'));
 			for (const name of Object.keys(this.commands)) {
-				lines.push(`    ${highlight(name)} => ${highlight(this.commands[name].name)}`);
+				const cmd = this.commands[name];
+				lines.push(`    ${highlight(name)} => ${highlight(cmd.name)}${cmd.modulePath ? note(' (import pending)') : ''}`);
 			}
 		}