+++ test/examples/external-module/extmod.js
@@ -1,5 +1,8 @@
-const CLI = require('../../../dist/index').default;
-const path = require('path');
+import CLI from '../../../src/index.js';
+import path from 'path';
+import { fileURLToPath } from 'url';
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 new CLI({
 	extensions: [ path.resolve(__dirname, '..', '..', 'fixtures', 'simple-module') ]