+++ test/test-actions.js
@@ -1,6 +1,10 @@
-import CLI, { Terminal } from '../dist/index';
+import CLI, { Terminal } from '../src/index.js';
 import path from 'path';
+import { expect } from 'chai';
+import { fileURLToPath } from 'url';
 import { WritableStream } from 'memory-streams';
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 describe('Actions', () => {
 	it('should emit an action to extensions', async function () {